Move TeardownInstallationSourceImageTask to payload
This task is specific for live_image payload so move it under its
folder.
