+++ pyanaconda/modules/payload/live/initialization.py
@@ -19,11 +19,6 @@ import os
 import glob
 
 from pyanaconda.modules.common.task import Task
-from pyanaconda.modules.payload.live.utils import get_local_image_path_from_url, \
-    url_target_is_tarfile
-from pyanaconda.payload.utils import unmount
-from pyanaconda.core.constants import IMAGE_DIR
-
 from pyanaconda.core.util import execWithRedirect
 
 from pyanaconda.anaconda_loggers import get_module_logger
@@ -68,44 +63,6 @@ class DownloadProgress(object):
         """Download complete."""
         self.report("Downloading image %(url)s (%(pct)d%%)" %
                     {"url": self.url, "pct": 100})
-
-
-class TeardownInstallationSourceImageTask(Task):
-    """Task to tear down installation source image."""
-
-    def __init__(self, image_path, url, image_mount_point):
-        """Create a new task.
-
-        :param image_path: destination path for image download
-        :type image_path: str
-        :param url: installation source image url
-        :type url: str
-        :param image_mount_point: Mount point of the source image
-        :type image_mount_point: str
-        """
-        super().__init__()
-        self._image_path = image_path
-        self._url = url
-        self._image_mount_point = image_mount_point
-
-    @property
-    def name(self):
-        return "Tear down installation source image."""
-
-    def run(self):
-        """Run tear down of installation source image."""
-        if not url_target_is_tarfile(self._url):
-            unmount(self._image_mount_point, raise_exc=True)
-            #FIXME: Payload and LiveOS stuff
-            # FIXME: do we need a task for this?
-            if os.path.exists(IMAGE_DIR + "/LiveOS"):
-                # FIXME: catch and pass the exception
-                unmount(IMAGE_DIR, raise_exc=True)
-                os.rmdir(IMAGE_DIR)
-
-        if not get_local_image_path_from_url(self._url):
-            if os.path.exists(self._image_path):
-                os.unlink(self._image_path)
 
 
 class UpdateBLSConfigurationTask(Task):
