+++ pyanaconda/modules/payload/payloads/live_image/live_image.py
@@ -33,9 +33,9 @@ from pyanaconda.modules.payload.base.initialization import CopyDriverDisksFilesT
 from pyanaconda.modules.payload.payloads.live_image.live_image_interface import \
     LiveImageHandlerInterface
 from pyanaconda.modules.payload.payloads.live_image.initialization import \
-    CheckInstallationSourceImageTask, SetupInstallationSourceImageTask
-from pyanaconda.modules.payload.live.initialization import UpdateBLSConfigurationTask, \
+    CheckInstallationSourceImageTask, SetupInstallationSourceImageTask, \
     TeardownInstallationSourceImageTask
+from pyanaconda.modules.payload.live.initialization import UpdateBLSConfigurationTask
 from pyanaconda.modules.payload.live.utils import get_kernel_version_list, \
     get_kernel_version_list_from_tar, url_target_is_tarfile
 from pyanaconda.modules.payload.live.installation import InstallFromImageTask, InstallFromTarTask
