+++ tests/nosetests/pyanaconda_tests/module_payload_live_image_test.py
@@ -31,9 +31,9 @@ from pyanaconda.modules.payload.payloads.live_image.live_image import LiveImageH
 from pyanaconda.modules.payload.payloads.live_image.live_image_interface import \
     LiveImageHandlerInterface
 from pyanaconda.modules.payload.payloads.live_image.initialization import \
-    CheckInstallationSourceImageTask, SetupInstallationSourceImageTask
-from pyanaconda.modules.payload.live.initialization import UpdateBLSConfigurationTask, \
+    CheckInstallationSourceImageTask, SetupInstallationSourceImageTask, \
     TeardownInstallationSourceImageTask
+from pyanaconda.modules.payload.live.initialization import UpdateBLSConfigurationTask
 from pyanaconda.modules.payload.base.initialization import CopyDriverDisksFilesTask
 from pyanaconda.modules.payload.live.installation import InstallFromImageTask, InstallFromTarTask
